New map and new starting point
camera is on starting point, and zergs makes his respawns on the north
side of the map

+++ Motor2D/Render.h
@@ -61,7 +61,7 @@ public:
 	bool			transition_active;
 	iPoint			dest_pos;
 	iPoint			middle_pos;
-	float			lerp_value;
+	int			lerp_value;
 };
 
 #endif // __RENDER_H__
